refactor(arrayConfiguration): move arrayConfiguration to shared folder
